Solves suspended vs buried issue
The original script counts buried cards (queue in -2,-3) along with the suspended ones (queue = -1).

This meant that most often than not, the suspended cards count included cards that are just hidden until the next day (hidden because they are related to cards you've just seen).

I've separated the counts, and also added a row for the buried cards at the stats table.